Further updates to file structure
