+++ finite-state-machine/state_machine.py
