+++ finite-state-machine/state_machine_3.py
